using common pkg for Namespace detection
