+++ pkg/util/kubernetes/apiserver/apiserver.go
@@ -10,7 +10,6 @@ package apiserver
 import (
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"strings"
 	"sync"
 	"time"
@@ -25,6 +24,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/util/cache"
+	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/apiserver/common"
 	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/hpa"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/util/retry"
@@ -256,7 +256,7 @@ func (c *APIClient) ComponentStatuses() (*v1.ComponentStatusList, error) {
 
 // GetTokenFromConfigmap returns the value of the `tokenValue` from the `tokenKey` in the ConfigMap `configMapDCAToken` if its timestamp is less than tokenTimeout old.
 func (c *APIClient) GetTokenFromConfigmap(token string, tokenTimeout int64) (string, bool, error) {
-	namespace := GetResourcesNamespace()
+	namespace := common.GetResourcesNamespace()
 	tokenConfigMap, err := c.Cl.CoreV1().ConfigMaps(namespace).Get(configMapDCAToken, metav1.GetOptions{})
 	if err != nil {
 		log.Debugf("Could not find the ConfigMap %s: %s", configMapDCAToken, err.Error())
@@ -298,7 +298,7 @@ func (c *APIClient) GetTokenFromConfigmap(token string, tokenTimeout int64) (str
 // UpdateTokenInConfigmap updates the value of the `tokenValue` from the `tokenKey` and
 // sets its collected timestamp in the ConfigMap `configmaptokendca`
 func (c *APIClient) UpdateTokenInConfigmap(token, tokenValue string) error {
-	namespace := GetResourcesNamespace()
+	namespace := common.GetResourcesNamespace()
 	tokenConfigMap, err := c.Cl.CoreV1().ConfigMaps(namespace).Get(configMapDCAToken, metav1.GetOptions{})
 	if err != nil {
 		return err
@@ -389,22 +389,6 @@ func getNodeList(cl *APIClient) ([]v1.Node, error) {
 		return nil, err
 	}
 	return nodes.Items, nil
-}
-
-// GetResourcesNamespace is used to fetch the namespace of the resources used by the Kubernetes check (e.g. Leader Election, Event collection).
-func GetResourcesNamespace() string {
-	namespace := config.Datadog.GetString("kube_resources_namespace")
-	if namespace != "" {
-		return namespace
-	}
-	log.Debugf("No configured namespace for the resource, fetching from the current context")
-	namespacePath := "/var/run/secrets/kubernetes.io/serviceaccount/namespace"
-	val, e := ioutil.ReadFile(namespacePath)
-	if e == nil && val != nil {
-		return string(val)
-	}
-	log.Errorf("There was an error fetching the namespace from the context, using default")
-	return "default"
 }
 
 // GetRESTObject allows to retrive a custom resource from the APIserver
